requirements.txt
wand
opencv-python
tqdm
pillow
pytesseract
cnstd
pix2tex
